+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Unreleased
 
+- Fix typo batching stats, changing send_items_in_success -> sent_items_in_success and send_items_in_error -> sent_items_in_error
+
 # 0.23.1 - 2024-04-18
 
 - Fix winston logger, remove colors